feat: add check for existing user

+++ src/handlers/roomHandler.ts
@@ -1,16 +1,17 @@
 import { IAddPlayerReq, ICreateGameRes, ICreateRoomReq, IUpdateRoomRes } from '../models/index.js';
-import { addPlayerToRoom, createRoom, createGame, deleteRoom } from '../dbHelpers/index.js';
+import { addPlayerToRoom, createRoom, createGame, deleteRoom, userAlreadyInRoom } from '../dbHelpers/index.js';
 import updateRoomsRes from './updateRooms.js';
 
 function roomHandler(
   req: ICreateRoomReq | IAddPlayerReq,
 ): { type: 'create_game'; resArr: ICreateGameRes[] } | IUpdateRoomRes {
-  if (req.type === 'create_room') {
-    createRoom();
+  const userId = req.id;
+  if (req.type === 'create_room' && !userAlreadyInRoom(userId)) {
+    createRoom(userId);
   }
-  if (req.type === 'add_user_to_room') {
+  if (req.type === 'add_user_to_room' && !userAlreadyInRoom(userId)) {
     const { indexRoom } = req.data;
-    const room = addPlayerToRoom(indexRoom, req.id);
+    const room = addPlayerToRoom(indexRoom, userId);
     if (room.roomUsers.length >= 2) {
       deleteRoom(indexRoom);
       const game = createGame(room.roomUsers);
@@ -21,7 +22,7 @@ function roomHandler(
             idGame: game.gameId,
             idPlayer: index,
           },
-          id: index,
+          id: 0,
         }),
       );
       return {
